chore: refactor ColorizeNumbers into components/common
