+++ src/components/optimizerForm/character/conditionals/DisplayFullPassives.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Flex } from 'antd';
-import ColorizeNumbers from 'components/optimizerForm/character/utils/ColorizeNumbers';
+import ColorizeNumbers from 'components/common/ColorizeNumbers';
 
 
 const FullPassives = (content: { [key:string]: { [key: string]: string}}) => {
